Move release:gen-payload command to a separate file
